Created Home Page
This website was originally intended to be about rice but that was an incredibly dull idea. I changed it to be a website on the Netflix anime, Great Pretender.

So far, I've created the home page and added most of the content that I wanted. All that's left to do is to work on the responsiveness as well as create a menu for mobile.

I also want to add some transitions when scrolling through the page.