Responsive Home Page
The home page is now responsive. When the display width is less than 1024px, the  navbar will be hidden and replaced with a hamburger menu.

The hamburger menu button has a slide animation and will also add an overlay for contrast.

I also created a css file specifically for the index page so the main.css file will look less messy.