Redesigned Home Page
The colours of the previous design seemed a bit weird so I changed up the colour palette to a dark-themed one.

I also made the nav bar fixed to the top of the screen. The background of the nav bar is supposed to change between transparent and coloured after scrolling past a certain point but I'm having a little trouble doing that.