Completed first section of Cases page
I decided that each section in the Cases page will consist of 3 parts.

One part will contain a synopsis, location and target of the cases/arc in the show.

The second part is the overview/plan of the scheme.

The third and last section are the important side-characters in the case/arc.

Each of these sections will have a different coloured background to distinguish between the sections within the section.

I also created a CSS class called 'case-card' along with the components within it which looks like a polaroid picture with a pin at the top.