Can select characters on characters page
Using the element's id and indexing the JSON with that id, I am able to retrieve the data needed for the elements in the character-info div. So upon pressing one of the character icons, the description changes to match the selected character.

Still need to edit the JSON file, add images for each of the characters and then fix the sizes of the images.