Fixes for tablet view
On the iPad Air (landscape) the PC nav bar was visible but couldn't interact with the cases <a> tag because it requires a mousehover in order to reveal  the dropdown. I added a click event listener on JS so it is accessible from a tablet.

The background also stretched because I was using the banner class but it was only on the cases stylesheet so I moved it over to the main stylesheet that's shared between all the webpages.