Fixed Cases Dropdown on Tablet
Before I added a toggle feature so that when the user clicked on Cases, it will show/hide the dropdown but it only hides the dropdown when the user clicks that button again.

I changed it so that it will always detect a click, and if the element that is clicked is the Cases link then it will show the dropdown. Otherwise, if the dropdown is visible and the user clicked somewhere else on the DOM, it will close the dropdown.